Use URL encoding to serialize data in SASS
To pass data from SASS to Javascript, the map-serialize function
generated JSON which libsass fails to properly interpolate when
compiling to CSS.

Using URL encoding instead avoids this issue and can serialize key/value
pairs in SASS with more stability.

On the Javascript side, the "parseQueryString" method is added to parse
this data back into a JS object. This method is adapted from
https://github.com/sindresorhus/query-string.